move InvalidAuth handling out of sensor
